backend/pyproject.toml
[tool.poetry]
name = "audio-assistant-backend"
version = "0.1.0"
description = ""
authors = ["shavs2077 <salva.neva96@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
faster-whisper = "^1.1.0"
transformers = "^4.47.1"
fastapi = {extras = ["standard"], version = "^0.115.6"}
python-multipart = "^0.0.20"
pyaudio = "^0.2.14"
llvmlite = "^0.43.0"
numba = "^0.60.0"
openai-whisper = "^20240930"
ollama = "^0.4.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
